SUEDE: move class
