+++ base/src/main/java/org/aya/core/serde/CompiledAya.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.library;
+package org.aya.core.serde;
 
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.DynamicSeq;
@@ -16,9 +16,6 @@ import org.aya.concrete.stmt.BindBlock;
 import org.aya.core.def.DataDef;
 import org.aya.core.def.Def;
 import org.aya.core.def.StructDef;
-import org.aya.core.serde.SerDef;
-import org.aya.core.serde.SerTerm;
-import org.aya.core.serde.Serializer;
 import org.aya.util.binop.OpDecl;
 import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
